+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Unreleased
+- Upgrade dependencies
+([#57](https://github.com/fs/react-base/pull/57))
+- Add json loader. Cleanup webpack configs. Change webpack logs mode
+([#57](https://github.com/fs/react-base/pull/57))
 - Use react entities as self imported
 ([#56](https://github.com/fs/react-base/pull/56))